+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 1.0.9
 
-* TODO: NumUtil updates(add subtract multiply divide without loosing precision). 
+* TODO: NumUtil updates(add subtract multiply divide without loosing precision). MoneyUtil precise conversion.
 
 ## 1.0.8